PR suggestion fixes. Extract work manager handling to the global TestingUtils file, so it can be used everywhere. Simplify/beautify code.
refs: MBL-19298
affects: Student
release note:
