+++ apps/student/build.gradle
@@ -233,7 +233,7 @@ dependencies {
     /* Android Test Dependencies */
     androidTestImplementation project(path: ':espresso')
     androidTestImplementation project(':dataseedingapi')
-    androidTestImplementation "androidx.work:work-testing:2.9.1" // Added this line
+    androidTestImplementation Libs.ANDROIDX_WORK_TEST
 
     /* Unit Test Dependencies */
     testImplementation Libs.JUNIT
