+++ automation/espresso/build.gradle
@@ -91,7 +91,7 @@ dependencies {
 
     implementation project(':dataseedingapi')
     implementation project(':login-api-2')
-    implementation 'androidx.work:work-testing:2.9.1'
+    implementation Libs.ANDROIDX_WORK_TEST
     androidTestImplementation Libs.COMPOSE_UI_TEST
 
     implementation project(':pandautils')
